backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-decouple==3.8
aiofiles==23.2.1
pillow==10.1.0
ffmpeg-python==0.2.0
redis==5.0.1
celery==5.3.4
boto3==1.34.0
opencv-python==4.8.1.78
moviepy==1.0.3
pydantic==2.5.0
httpx==0.25.2